moves update modal into it's own component
